Move random number generation utilities to random.py
